Medibork changes
This will change the load out for mediborks. 
It will add:

****Medical Gripper. Allows borgs to pick up beakers and open bottles. As well as pick up phoron crystals for grinding. This will allow them to create phoron based chems and setup cryo byond (heh) just turning them on. It also allows them to pick up pills and force them on patients such as carbon pills. It also allows picking up blood packs and placing them in IV's.
Full list of items are that can be picked up are:
 /obj/item/weapon/reagent_containers/glass,
		/obj/item/weapon/storage/pill_bottle,
		/obj/item/weapon/reagent_containers/pill,
		/obj/item/weapon/reagent_containers/blood,
                /obj/item/stack/material/phoron

****Traumakits, Advance Burn kits, and Splints.

Items work just as the normal ones. The only twist is that they got practically infinite ammo.  (1000 actually according to the code) 
However, Do not be fooled by the amount of charge it has when examining it.
But they may only apply it once per-limb so no heal spamming. Even after that limb takes damage again. You got to wait until the "already treated" thing wears off to reapply.

*Traumakits: The main use of the trauma kit is to stop external bleeding.
*Burn Pack: Main use of burnpack (ointment) Is to treat burn damage overtime. It seems useless next to the hypospray but who knows. It treats about 2.2 burn damage on use so it just might bring a burned victim's damage down enough to be defibbed.
*Splints: Lets you treat a patient with bone damage as best you can while waiting for a surgeon to show up. Of course won't actually fix anything, but at least they'll stop falling down in pain or dealing self damage by moving with broken bones.