Nerf HoS Armored Coat
The HoS armored coat was too powerful compared to other armors. The numbers have been brought down a little, but are still slightly superior to standard security vests. The head of security vest, however, was only slightly nerfed with melee being reduced by five. It keeps its superior bullet and energy protection, but bomb protection is being reduced to 15 to be in line with regular security vest.