Door tweak 4
Changed all the command doors again to remove beeping and to shorten sound to match animations a little bitter. Updated Vault Door sound effects because one of them was broken and I forget which one, so have both. Also updated the Secure door close because that too was broken, and made it a little longer to match the animation.