New Job Title for Doctor "Resleeving Technician"
Added a new job title for the doctor role, the Resleeving Technician. It matches the Nurse role in the character's uniform appearance, but the hope is that anyone who picks the role will instead be focusing on resleeving people and keeping a close eye on the crew health monitor. Suggested by Decker.