Added EVA RIG to map + CentCom Reagents
THIS UPDATE IS NOW GOOD TO GO!

The KWI Rig has been added to the map. Two in EVA, one in emergency EVA replacing one of the spacesuits, and one in the NJ's admeme room.

ALSO, on the side, Mira asked for one of the chem dispensers at central to dispense reagents instead of chems. So you have asked, so you shall receive.