Bye-bye, Mechs. Hello, delicious points!
Comments out requests for any sort of mechs and makes the points rewards randomised in proportion to the event's severity. Minor would give you a low amount of points. Major would give ya a heaping pile of points. Minor edit to the paper thing so it's not blending with another word.