Solarmoth Code Addition
Adds the Solarmoth to the game finally. Adds code for the Solargrub to also track an how much power it has arbitrarily drained. Can possibly be used for other things as well. A global variable has been added in the Solarmoth code that the Solargrub keeps track of to see how many Solarmoths there has been in the world. Only up to two moths can ever be spawned in the world before the rest of the grubs decide to no longer evolve.