Teshari RIG sprite fixes
Had to fix naming conventions and do a little sprite artwork.

Had to rename sprites and do sprite artwork.
Changed naming convention from rig-whatever to whatever_rig and whatever_rig_sealed. This will fix the broken teshari RIG sprites on the Polaris map that we currently use.

Also, modeled all of the other department RIG helmets after the CE RIG helmet for the Teshari. Because it looked so freaking cool.