Dorm Atmos Overhaul
Reworks the air supply pipes in an attempt to prevent spiders from ever appearing inside the dorms so you can scene in peace without worry of getting the drop on ya from a hunter. Also adds AR-B glasses to CD Office, plain AR glasses to HOP, and removes useless timeclocks in and around the bar.