Timeclocks Are Back!
Adding this to stay mostly synched to another PR I have open. Likely conflicts may occur but be restricted to the trade shuttle. This map-edit includes the addition of timeclocks throughout the station rather than the previous limitation of the bar area. It also removes Honk Jr. as well as adds spawnpoints which is used in said other PR.