Pay equality
Apparently humans got an economic modifier of 10. Unathi and a few other races had a modifier of 7. Sergals and a whole slew of other sepecies had a modifier of 2, which is really low and can affect the amount of money you get.

We don't believe in discrimination here. All pay equalized to economic modifier 10.

Note that this relates to the amount of money you spawn in with. I felt like doing this just cuz, but it will also play well into the future casino event!