Some carp spawns moved inside
Some carp spawns have been moved to inside the station. 8 are in the caves around the station, 7 are on the vents near the outer-edges of the station.

The carp have been finding ways of squeezing in through the rocks, making their way into the caves of the station. In said caves they have also found ways into the ventilation system.

Have security watch out for them!