changes to space vines, additional ranged mob behaviour and three plant mobs
Thanks to Razgriz and Dragonhark for coding most of the three plant mobs; piranha plant, spitter and pitcher plant, and Skits for providing the sprites.

I have with help finished the mobs with spitter stunning victims for other plants to eat, pitcher heals when having a meal and both it and spitter produces soyelent which can have fture possible uses!
If the stars align and a space vine has the carnivorous trait and enough potency, it has a chance to spawn these mobs when spreading!
additionally theres now two new variables, one that makes ranged mob stop targeting unconscious victims and another that makes mobs immune to entangling and attacking vines!